+++ CHANGES.md
@@ -13,6 +13,10 @@
   This makes prefixstr/suffixstr (`**>`/`<**`) redundant, they are removed.
 * The various list combinators now accept a tyregex as separator.
   The old behavior can be recovered by combining with `Tyre.str`.
+* Add the `start` and `stop` combinators.
+* The ~whole argument for compile and route is removed.
+  tyregex don't match the whole string by default anymore.
+  You can use `Tyre.whole_string` or `Tyre.start` and `Tyre.stop` instead.
 
 ## 0.1.1 (09 September 2016)
 * Fix a bug with nested repetitions. Also avoid some copying of the original string.